hores decimals

+++ app/src/main/java/com/pes/become/backend/adapters/DomainAdapter.java
@@ -372,10 +372,10 @@ public class DomainAdapter {
      * Metode per obtenir les hores dedicades a cada tema en la rutina seleccionada
      * @return ArrayList on a cada posicio hi han les hores dedicades al tema equivalent a la posicio (i.e. a la posicio 0 hi ha les hores dedicades a Music)
      */
-    public ArrayList<Integer> getHoursByTheme(){
-        ArrayList<Integer> hoursByTheme = new ArrayList<>();
+    public ArrayList<Double> getHoursByTheme(){
+        ArrayList<Double> hoursByTheme = new ArrayList<>();
         for(int theme=0; theme<Theme.values().length; ++theme){
-            int hours = currentUser.getHoursByTheme(Theme.values()[theme]);
+            double hours = currentUser.getHoursByTheme(Theme.values()[theme]);
             hoursByTheme.add(theme, hours);
         }
         return hoursByTheme;
